Revise `protocol/progress.go` and update its documentation

+++ cmd/client/main.go
@@ -292,7 +292,7 @@ func transferFile(ctx context.Context, conn net.Conn, filePath string, relPath .
 	startTime := time.Now()
 
 	// Create a progress reader to track upload progress.
-	progressReader := protocol.NewProgressReader(file, int64(header.FileSize), "Uploading")
+	progressReader := protocol.NewProgressReader(file, header.FileSize, "Uploading")
 
 	// Create a context-aware writer that can be interrupted during shutdown.
 	ctxWriter := &contextWriter{
